+++ processor/src/test/resources/bad_input/com/example/dependency/ThrowsDependency.java
@@ -9,7 +9,7 @@ import java.io.IOException;
 public abstract class ThrowsDependency
 {
   @Dependency
-  DisposeTrackable getTime()
+  final DisposeTrackable getTime()
     throws IOException
   {
     return null;